requirements.txt
streamlit
google-generativeai
youtube-search
ics